backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-extensions
django-cors-headers
requests
google-auth